Implement new TimedOrderedLogger
With a very high rate of messages, the OrderedLogger sometimes had the
problem, that even the 10000 message limit didn't stop unordered
messages from being logged. So TimedOrderedLogger instead determines the
timespan currently in the buffer, and uses that to decide whether a
commit to the file should happen. It also doesn't use any threading for
now, to keep it as simple and stupid as possible.